Ajustement du controler d'enregistrement d'utilisateur pour que les erreurs du service passe plus directement au retour. Changement d'emplacmeent pour les helpers.ts
