+++ Dockerfile.dev
@@ -5,7 +5,7 @@ RUN apk --no-cache add --virtual .builds-deps build-base python3
 
 ## API
 WORKDIR /api
-COPY ["package.json", "package-lock.json", "./"]
+COPY ["package.json", "./"]
 
 
 RUN npm install
