+++ src/Authentification/Controllers/RegistrationController.ts
@@ -2,7 +2,7 @@ import { ReasonPhrases, StatusCodes } from "http-status-codes";
 import { ApiResponseContract } from "../../Http/Responses/ApiResponse";
 import { ErrorResponse } from "../../Http/Responses/ErrorResponse";
 import LogHelper from "../../Monitoring/Helpers/LogHelper";
-import {UsersService, User, UserDocument, UserValidation} from "../../Users/UsersDomain";
+import {UsersService, User, UserDocument, UserValidation, UserSchema} from "../../Users/UsersDomain";
 import config from "../../config";
 
 
@@ -29,31 +29,30 @@ export class RegistrationController {
         }
 
         const formattedData = this.formatRequestDataForDocument(requestData);
-
         const createdDocumentResponse:ApiResponseContract = await this.service.insert(formattedData);
-        LogHelper.debug("userController", createdDocumentResponse);
 
-        if (createdDocumentResponse !== undefined)
-            return createdDocumentResponse;
 
-        return ErrorResponse.create(
-            new Error(ReasonPhrases.INTERNAL_SERVER_ERROR),
-            StatusCodes.INTERNAL_SERVER_ERROR,
-            'Les données semblent être ok, mais la création n\'a pas eu lieu.'
-        );
+        if (!createdDocumentResponse.error) {
+            createdDocumentResponse.data = UserSchema.dataTransfertObject(createdDocumentResponse.data);
+            //generate un token ? direct ?
+            // ou on fait connecter l'utilisateur ?
+            // Il faut faire le flow post inscription.
+        }
+        return createdDocumentResponse;
     }
 
     public validateData(data:any):any
     {
         let message = "validateData register";
-
+/*
+//pour appliquer la validation sur les données reçus.
         for (let property in UserValidation) {
             const rules:Array<string> = UserValidation[property];
             for (let property of rules) {
 
             }
         }
-
+*/
         let isValid = typeof data === 'object';
         isValid = data.name !== undefined && isValid;
         isValid = data.username !== undefined && isValid;
