+++ src/Helpers/DateTime.ts
@@ -1,0 +1,9 @@
+//TIME Helpers
+
+/**
+ * Return a now date object directly
+ * Create a new instance of Date every time.
+ */
+export const now = () => {
+    return new Date().getTime();
+}
